Update install-service.sh
Added initial yn question

+++ install-service.sh
@@ -1,11 +1,19 @@
-sudo adduser --system jamulus
-sudo adduser jamulus www-data
+echo -n "JAMULUS SERVICE INSTALLATION. Are you sure? This script will overwrite the /etc/systemd/system/jamulus.service, if it exists. (y/n)? "
+read answer
+if [ "$answer" != "${answer#[Yy]}" ] ;then
 
-sudo cp files/jamulus.service /etc/systemd/system/jamulus.service
-sudo chmod 644 /etc/systemd/system/jamulus.service
+  sudo adduser --system jamulus
+  sudo adduser jamulus www-data
 
-sudo systemctl daemon-reload
+  sudo cp files/jamulus.service /etc/systemd/system/jamulus.service
+  sudo chmod 644 /etc/systemd/system/jamulus.service
 
-sudo systemctl start jamulus
-sudo systemctl enable jamulus
-sudo systemctl status jamulus
+  sudo systemctl daemon-reload
+
+  sudo systemctl start jamulus
+  sudo systemctl enable jamulus
+  sudo systemctl status jamulus
+
+else
+    echo -n "JAMULUS SERVICE INSTALLATION canceled."
+fi
